refactor: update page router
